Moved SERVICE_NAME to separate module for easy import
