+++ services/fis/src/fis/adapters/outbound/vault/client.py
@@ -24,7 +24,7 @@ from hvac.api.auth_methods import Kubernetes
 from pydantic import Field, SecretStr, field_validator
 from pydantic_settings import BaseSettings
 
-from fis.config import SERVICE_NAME
+from fis.constants import SERVICE_NAME
 from fis.ports.outbound.vault.client import VaultAdapterPort
 
 tracer = SpanTracer(SERVICE_NAME)
